agent/requirements.txt
livekit==0.17.0
python-dotenv==1.0.0
requests==2.31.0
